requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements --dev
#

# Note: in pipenv 2020.x, "--dev" changed to emit both default and development
# requirements. To emit only development requirements, pass "--dev-only".

-i https://pypi.org/simple
aiohttp==3.8.5; python_version >= '3.6'
aiosignal==1.3.1; python_version >= '3.7'
asgiref==3.7.2; python_version >= '3.7'
async-timeout==4.0.2; python_version >= '3.6'
attrs==23.1.0; python_version >= '3.7'
cachetools==5.3.1; python_version >= '3.7'
certifi==2023.5.7; python_version >= '3.6'
charset-normalizer==3.2.0; python_version >= '3.7'
django-cors-headers==4.2.0
django-debug-toolbar==4.1.0
django-environ==0.10.0
django-extensions==3.2.3
django-storages[google]==1.13.2
django-widget-tweaks==1.4.12
django==4.2.3; python_version >= '3.8'
frozenlist==1.4.0; python_version >= '3.8'
google-api-core==2.11.1; python_version >= '3.7'
google-auth==2.22.0; python_version >= '3.6'
google-cloud-appengine-logging==1.3.1; python_version >= '3.7'
google-cloud-audit-log==0.2.5; python_version >= '3.7'
google-cloud-core==2.3.3; python_version >= '3.7'
google-cloud-logging==3.6.0
google-cloud-secret-manager==2.16.2
google-cloud-storage==2.10.0
google-crc32c==1.5.0; python_version >= '3.7'
google-resumable-media==2.5.0; python_version >= '3.7'
googleapis-common-protos==1.59.1; python_version >= '3.7'
grpc-google-iam-v1==0.12.6; python_version >= '3.7'
grpcio-status==1.56.2
grpcio==1.56.2
gunicorn==21.2.0
idna==3.4; python_version >= '3.5'
multidict==6.0.4; python_version >= '3.7'
openai==0.27.8
packaging==23.1; python_version >= '3.7'
pillow==10.0.0
proto-plus==1.22.3; python_version >= '3.6'
protobuf==4.23.4; python_version >= '3.7'
psycopg2-binary==2.9.6
pyasn1-modules==0.3.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
pyasn1==0.5.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
regex==2023.6.3; python_version >= '3.6'
requests==2.31.0; python_version >= '3.7'
rsa==4.9; python_version >= '3.6' and python_version < '4'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sqlparse==0.4.4; python_version >= '3.5'
tiktoken==0.4.0
tqdm==4.65.0; python_version >= '3.7'
typing-extensions==4.7.1; python_version < '3.11'
urllib3==1.26.16; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
yarl==1.9.2; python_version >= '3.7'
